[core] Extract the HDF5 format into heat/core/io/hdf5.py
`supports_hdf5`, `load_hdf5`, `save_hdf5`, `load_multiple_hdf5` and the
`DNDarray.save_hdf5` binding move out of io.py, completing the split:
io.py is now just `load`, `save` and `size_from_slice`.

`supports_hdf5` is built on `_config.is_installed` and h5py is resolved
inside the functions, so it is no longer imported at `import heat` time.
The inline `h5py.get_config().mpi` checks become
`_config.hdf5_has_mpi()`, and the "falling back to serial I/O" warning
moves from import time to first use, emitted once on rank 0.

The write-mode check now goes through `utils.sanitize_write_mode`.
